**update v1.55 support for mythicmobs 4.9.0**

+++ src/main/java/com/gmail/berndivader/mythicmobsext/events/MythicMobsExtItemDropEvent.java
@@ -39,7 +39,7 @@ Cancellable {
         return this.am.getEntity().getBukkitEntity();
     }
 
-    public int getMobLevel() {
+    public double getMobLevel() {
         return this.am.getLevel();
     }
